feat: add Redis, move cassandra package
