+++ go.mod
@@ -1,14 +1,17 @@
-module github.com/EdanStasiuk/LiteCode/apps/backend/server
+module github.com/EdanStasiuk/LiteCode
 
 go 1.24.5
 
 require (
 	github.com/gin-gonic/gin v1.10.1
 	github.com/gocarina/gocsv v0.0.0-20240520201108-78e41c74b4b1
+	github.com/redis/go-redis/v9 v9.12.1
 	gorm.io/gorm v1.25.10
 )
 
 require (
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/golang/snappy v0.0.3 // indirect
 	github.com/hailocab/go-hostpool v0.0.0-20160125115350-e80d13ce29ed // indirect
 	github.com/jackc/pgpassfile v1.0.0 // indirect
