+++ pkg/cassandra/session.go
